qt/windows: hide CMD window when queriyng for darkmode
When starting the app and when toggling darkmode, a CMD window would
pop up on Windows very quickly during the execution of `reg query
...`.

Had to split it into separate files with build-tags as the
`HideWindow` struct field does not exist only on Windows and not on
the other OS.
