+++ frontends/qt/server/detectdarktheme_darwin.go
@@ -1,0 +1,39 @@
+// Copyright 2023 Shift Crypto AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build darwin
+// +build darwin
+
+package main
+
+import (
+	"os/exec"
+	"strings"
+
+	"github.com/digitalbitbox/bitbox-wallet-app/util/logging"
+)
+
+// detect theme used by OS and return true if it's dark
+func detectDarkTheme() bool {
+	log := logging.Get().WithGroup("server")
+	cmd := exec.Command("defaults", "read", "-g", "AppleInterfaceStyle")
+	out, err := cmd.Output()
+	if err == nil {
+		log.Info("MacOS theme: " + string(out))
+		if strings.TrimSpace(string(out)) == "Dark" {
+			return true
+		}
+	}
+	return false
+}
