+++ frontends/qt/server/detectdarktheme_windows.go
@@ -1,0 +1,43 @@
+// Copyright 2023 Shift Crypto AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build windows
+// +build windows
+
+package main
+
+import (
+	"os/exec"
+	"strings"
+	"syscall"
+
+	"github.com/digitalbitbox/bitbox-wallet-app/util/logging"
+)
+
+// detect theme used by OS and return true if it's dark
+func detectDarkTheme() bool {
+	log := logging.Get().WithGroup("server")
+	const regKey = `HKCU\Software\Microsoft\Windows\CurrentVersion\Themes\Personalize`
+	const regName = `AppsUseLightTheme`
+	cmd := exec.Command("reg", "query", regKey, "/v", regName)
+	cmd.SysProcAttr = &syscall.SysProcAttr{HideWindow: true}
+	out, err := cmd.Output()
+	if err == nil {
+		log.Info("windows theme: " + string(out))
+		if strings.Contains(string(out), "0x0") {
+			return true
+		}
+	}
+	return false
+}
